New CPP TokenManager Generator
